% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmsa.R
\name{alpha_gene_systems}
\alias{alpha_gene_systems}
\title{Alpha panel gene -> system map (bundled)}
\usage{
alpha_gene_systems()
}
\value{
data.frame with system_id, system, gene, n_kept, usability
}
\description{
Read the Project Alpha gene codebook shipped with the package:
549 genes assigned to 30 biological systems.
}
\examples{
gs <- alpha_gene_systems()
dim(gs)
length(unique(gs$system_id))
head(gs[, c("system_id", "system", "gene", "n_kept", "usability")], 4)
}
