% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{dmsa_mvalues}
\alias{dmsa_mvalues}
\title{M-value transform with clamping}
\usage{
dmsa_mvalues(beta, eps = 1e-04)
}
\arguments{
\item{beta}{numeric matrix or vector of beta values in (0,1)}

\item{eps}{clamp bound}
}
\description{
M-value transform with clamping
}
